Migrate RBACHook to TS
