--- packages/utils/src/RBACHook/index.js
@@ -1,2 +1,0 @@
-export { default } from './RBACHook';
-export * from './RBACHook';
